+++ internal/web/html/templates/browser/layout.html
@@ -83,7 +83,18 @@
 
   <script src="/assets/js/libs/htmx.min.js"></script>
   <script src="/assets/js/libs/response-targets.js"></script>
-  <script type="module" src="/assets/js/sidenav.mjs"></script>
+
+  <script type="module">
+import {Sidenav, Datatable, Navbar, Dropdown, Modal, Select, initMDB} from "/assets/js/libs/mdb.es.min.js";
+import {SetupSideNav, SetupBoostrapElems} from "/assets/js/setup.mjs"
+
+initMDB({Sidenav, Datatable, Navbar, Dropdown, Modal, Select});
+SetupSideNav()
+SetupBoostrapElems()
+  </script>
+
+  <script type="module" src="/assets/js/file-upload.mjs"></script>
+
 </body>
 
 </html>
