+++ internal/web/html/templates/browser/page.html
@@ -53,6 +53,3 @@
 <input type="hidden" id="folder-path-meta" value="{{$.Folder.Path}}">
 <input type="hidden" id="space-id-meta" value="{{$.Folder.Space.ID}}">
 </section>
-
-<script type="module" src="/assets/js/file-upload.mjs"></script>
-<script type="module" src="/assets/js/sidenav.mjs"></script>
