+++ internal/web/html/templates/settings/layout.html
@@ -99,7 +99,17 @@
 
   <script src="/assets/js/libs/htmx.min.js"></script>
   <script src="/assets/js/libs/response-targets.js"></script>
-  <script type="module" src="/assets/js/sidenav.mjs"></script>
+<!--  <script type="module" src="/assets/js/sidenav.mjs"></script> -->
+
+  <script type="module">
+import {Sidenav, Datatable, Navbar, Dropdown, Modal, Select, initMDB} from "/assets/js/libs/mdb.es.min.js";
+import {SetupSideNav, SetupBoostrapElems} from "/assets/js/setup.mjs"
+
+initMDB({Sidenav, Datatable, Navbar, Dropdown, Modal, Select});
+SetupSideNav()
+SetupBoostrapElems()
+  </script>
+
 </body>
 
 </html>
